seccomp

Module raw

The raw FFI to libseccomp. This is all unsafe. Use it only if you understand what values should be returned, what objects you need to manage, etc.

Structs

seccomp_data

Syscall data.

seccomp_notif

A notification from the kernel.

seccomp_notif_resp

A Notification Response structure.

Enums

scmp_filter_attr

Attributes. ACT_DEFAULT is not included because seccomp_init already takes it.

Statics

SCMP_ACT_ALLOW

Allow the action.

SCMP_ACT_KILL_PROCESS

Kill the process

SCMP_ACT_KILL_THREAD

Kill the thread

SCMP_ACT_LOG

Log the request to the Audit log

SCMP_ACT_NOTIFY

Request a decision from the notify monitor.

SCMP_ACT_TRAP

Trap Signal

Functions

seccomp_api_get⚠

Get the current API level.

seccomp_arch_native⚠

Get the native architecture.

seccomp_attr_set⚠

Set an attribute. See seccomp.h for expected values, or see the Attributes trait.

seccomp_export_bpf⚠

Export the filter to BPF for Bubblewrap.

seccomp_init⚠

Initialize a SECCOMP context

seccomp_load⚠

Load the filter into the current process.

seccomp_notify_alloc⚠

Allocate a notification pair.

seccomp_notify_fd⚠

Get the Notify FD to receive and respond over.

seccomp_notify_free⚠

Free a notification pair.

seccomp_notify_id_valid⚠

Check if a event is still valid.

seccomp_notify_receive⚠

Receive an event from the kernel.

seccomp_notify_respond⚠

Send a response to an event.

seccomp_release⚠

Release a context.

seccomp_rule_add⚠

Add a rule.

seccomp_set_priority⚠

Set the priority of a syscall.

seccomp_syscall_resolve_name⚠

Resolve names, like “ptrace” to the associated number for the current architecture

seccomp_syscall_resolve_name_arch⚠

seccomp_syscall_resolve_num_arch⚠

Resolve a syscall number with an architecture to the name.

Type Aliases

scmp_filter_ctx

A SECCOMP context.