Crate seccomp

A wrapper for libseccomp.

Modules

action

Wrapper for SCMP_ACT.

attribute

A wrapper on SCMP_FLTATR.

filter

A wrapper around a SECCOMP context.

notify

A wrapper for the SECCOMP Notify interface.

raw

The raw FFI to libseccomp. This is all unsafe. Use it only if you understand what values should be returned, what objects you need to manage, etc.

syscall

Helper functions wrapper the seccomp crate.

Enums

Error

An error for all aspects of the SECCOMP crate.

Functions

get_architecture

Get the current architecture.